atv_demod: fix coverity issues(V4.05) [3/3]
PD#SWPL-217787

Problem:
fix coverity issues

Solution:
1.fix coverity issues
2.VER V4.05

Verify:
t6d

+++ drivers/media/atv_demod/atv_demod_access.c
@@ -114,7 +114,7 @@ int amlatvdemod_periphs_reg_write(unsigned int reg, unsigned int val)
 
 void atv_dmd_wr_reg(unsigned char block, unsigned char reg, unsigned long data)
 {
-	unsigned long reg_addr = (block << 8) + reg * 4;
+	unsigned long reg_addr = (unsigned long)(block << 8) + reg * 4;
 
 	amlatvdemod_reg_write(reg_addr, data);
 }
